bug: fix login broken because on condition check
