+++ packages/api/controller/auth/user.controller.js
@@ -1,4 +1,4 @@
-const { response } = require("../../utilities/response");
+const { response, authResponse } = require("../../utilities/response");
 const { getTokens } = require("./authThirdPartyController");
 const { register } = require("../../repository/user.repository");
 const { userCollection } = require("../../database/models/userSchema");
@@ -7,8 +7,8 @@ const {
   generateEmailVerificationLink,
   verifyLink,
 } = require("../../utilities/generateToken");
-const emailService = require("../../services/email.service"); 
-const {SIGNUP_TEMPLATE} = require('../../utilities/email.template') 
+const emailService = require("../../services/email.service");
+const { SIGNUP_TEMPLATE } = require("../../utilities/email.template");
 
 async function registerUser(req, res) {
   try {
@@ -34,7 +34,7 @@ async function registerUser(req, res) {
           );
 
     const checkEmailExist = await userCollection.findOne({ email });
-
+    console.log(checkEmailExist);
     if (checkEmailExist)
       return res
         .status(409)
@@ -55,7 +55,6 @@ async function registerUser(req, res) {
         action_url: verificationLink,
       },
     });
-
 
     if (!user)
       return res
@@ -79,56 +78,82 @@ async function registerUser(req, res) {
     );
   }
 }
+async function login(req, res) {
+  // retrieve the email and password
+  const { email, password } = req.body;
 
+  const user = await userCollection.findOne({ email });
+  
+  if (!user) {
+    return res
+      .status(401)
+      .json({ msg: `User with email ${email} does not exist.` });
+  }
+  
+  // comparing password
+  const validPassword = await user.comparePassword(password);
+
+  if (!validPassword) {
+    return res.status(401).json({ msg: "Invalid email or password" });
+  }
+
+  return res.status(200).json(
+    response({
+      success: true,
+      message: "User login successfully",
+      data: authResponse(user),
+    })
+  );
+}
 async function googleAuthUserSignUp(req, res) {
-	const { name, email } = await getTokens(req.query.code);
+  const { name, email } = await getTokens(req.query.code);
 
-	//Check if user already exist
-	const user = await userCollection.findOne({ email });
+  //Check if user already exist
+  const user = await userCollection.findOne({ email });
 
-	if (user) {
-		const data = {
-			_id: user._id,
-			firstname: user.firstName,
-			lastname: user.lastName,
-			username: user.username,
-			email: user.email,
-			language: user.language,
-			token: user.generateAuthToken(),
-		};
-		return res.status(200).json(
-			response({
-				success: true,
-				message: "User logged in Sucessfully",
-				data: data,
-			})
-		);
-	} else {
-		const randomUserCode = (Math.random() + 1).toString(36).substring(7);
-		const newName = name.split(" ");
+  if (user) {
+    const data = {
+      _id: user._id,
+      firstname: user.firstName,
+      lastname: user.lastName,
+      username: user.username,
+      email: user.email,
+      language: user.language,
+      token: user.generateAuthToken(),
+    };
+    return res.status(200).json(
+      response({
+        success: true,
+        message: "User logged in Sucessfully",
+        data: data,
+      })
+    );
+  } else {
+    const randomUserCode = (Math.random() + 1).toString(36).substring(7);
+    const newName = name.split(" ");
 
-		const data = {
-			firstName: newName[0],
-			lastName: newName[1],
-			email: email,
-			username: slugify(name) + randomUserCode,
-			password: "password",
-		};
-		const user = await register(data);
+    const data = {
+      firstName: newName[0],
+      lastName: newName[1],
+      email: email,
+      username: slugify(name) + randomUserCode,
+      password: "password",
+    };
+    const user = await register(data);
 
-		if (!user)
-			return res
-				.status(500)
-				.json(response({ success: false, message: "User not created" }));
+    if (!user)
+      return res
+        .status(500)
+        .json(response({ success: false, message: "User not created" }));
 
-		return res.status(201).json(
-			response({
-				success: true,
-				message: "User created successfully",
-				data: user,
-			})
-		);
-	}
+    return res.status(201).json(
+      response({
+        success: true,
+        message: "User created successfully",
+        data: user,
+      })
+    );
+  }
 }
 
 async function verifyMail(req, res) {
@@ -148,5 +173,4 @@ async function verifyMail(req, res) {
   }
 }
 
-module.exports = { registerUser, googleAuthUserSignUp, verifyMail };
-
+module.exports = { registerUser, googleAuthUserSignUp, verifyMail, login };
